Add files via upload
Hangman game created in Dr. Barbara Anthony's Computer Science I course. Credit to Dr. Anthony for the project and partner Nolan Winsman. Hangman is a game where the user must guess a random word correctly; otherwise, they lose lives in the form of a stick figure being drawn. This version of Hangman is unique in that it will cheat and keep changing the word being guessed until the user has ruled out 80% of the words available in the hangmanWords.txt file. This project demonstrated the importance of creating code that is robust to invalid user input.