Add files via upload
This is a game where the user must guess the numbers contained inside of a bag. At the start of a game, the user is asked to specify the range of the numbers and the quantity of numbers in the bag. With each guess, the user is told how many numbers they got correct but not which numbers. The user is allowed to guess until they have finally guessed the complete contents of the bag, meaning they cannot lose. This project demonstrated the power of the bag data structure in managing data in an unordered manner and the importance of creating robust code to avoid a failure of the program in the event a user enters an invalid entry. Credit to Dr. Jacob Schrum for the project idea and framework, Frank Carrano for the bag interface, and partner Kirby Steckel.