Add files via upload
This is a program that evaluates LISP expressions. This program accomplishes this by first converting the prefix LISP expression to post-fix format. The program then utilizes a stack to evaluate the expression. Credit to Dr.  Jacob Schrum for the template, Frank Carrano for the stack interface, and partner Amir Samiepour.