Add files via upload
This program searches for the longest common sequence in a strand of DNA or RNA. It utilizes a rudimentary approach ignoring many biological factors in DNA/RNA pairing. Furthermore, it includes methods demonstrating the effectiveness of memoization in speeding up sequence searching. Credit to Dr. Jacob Schrum of Southwestern University for the assignment template and partner Adrian Bonner.